A gcc fix in LinearizerBase.

+++ hermes2d/src/views/linearizer_base.cpp
@@ -138,7 +138,7 @@ namespace Hermes
           double mesh_size_x = top_right_x - bottom_left_x, mesh_size_y = top_right_y - bottom_left_y;
 
           double mesh_size_times_threshold = LinearizerBase::large_elements_fraction_of_mesh_size_threshold * mesh_size_x * mesh_size_y;
-          double ratio = (LIN_MAX_LEVEL * std::pow(element_area / mesh_size_times_threshold, 0.2)) * (std::sqrt(polynomial_order - 1));
+          double ratio = (LIN_MAX_LEVEL * std::pow(element_area / mesh_size_times_threshold, 0.2)) * (std::sqrt((double)polynomial_order - 1));
           this->level_map[e->id] = std::min<int>(LIN_MAX_LEVEL, (int)ratio);
         }
